Create zsync control file
Add a Makefile step to create a zsync control file. This is a small file
containing hashes of blocks that can be used by zsync to download a
delta based on a previously downloaded file. This approach works on any
HTTP server that supports download ranges. In the case of archiso it can
reduce data download by 40-50%, and will use any older image as a seed
file with no additional effort.